Move skull code to cheats module
